Split the cycle classes out of main adapter file, util methods
