+++ mangum/adapter.py
@@ -1,191 +1,15 @@
 import base64
 import asyncio
-import enum
-import os
 import traceback
 import urllib.parse
 import typing
 import json
-from dataclasses import dataclass, field
-
-import boto3
-import botocore
+from dataclasses import dataclass
 
 from mangum.lifespan import Lifespan
-from mangum.utils import get_logger
-from mangum.types import (
-    ASGIScope,
-    ASGIMessage,
-    ASGIApp,
-    AWSEvent,
-    AWSContext,
-    AWSResponse,
-)
-
-
-def make_response(content: str, status_code: int = 500):
-    return {
-        "statusCode": status_code,
-        "isBase64Encoded": False,
-        "headers": {"content-type": "text/plain; charset=utf-8"},
-        "body": content,
-    }
-
-
-def get_connections():
-    db = boto3.resource("dynamodb")
-    connections = db.Table(os.environ["TABLE_NAME"])
-    return connections
-
-
-def send_to_connections(*, data, connections, items, endpoint_url):
-    apigw_management = boto3.client(
-        "apigatewaymanagementapi", endpoint_url=endpoint_url
-    )
-    for item in items:
-        try:
-            apigw_management.post_to_connection(
-                ConnectionId=item["connectionId"], Data=data
-            )
-        except botocore.exceptions.ClientError as exc:
-            status_code = exc.response.get("ResponseMetadata", {}).get("HTTPStatusCode")
-            if status_code == 410:
-                connections.delete_item(Key={"connectionId": item["connectionId"]})
-            else:
-                return make_response("Connection error", status_code=500)
-    return make_response("OK", status_code=200)
-
-
-class ASGICycleState(enum.Enum):
-    REQUEST = enum.auto()
-    RESPONSE = enum.auto()
-
-
-@dataclass
-class ASGICycle:
-
-    scope: ASGIScope
-    state: ASGICycleState = ASGICycleState.REQUEST
-    binary: bool = False
-    response: dict = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        self.loop = asyncio.get_event_loop()
-        self.app_queue = asyncio.Queue(loop=self.loop)
-
-    def __call__(self, app: ASGIApp):
-        asgi_instance = app(self.scope, self.receive, self.send)
-        asgi_task = self.loop.create_task(asgi_instance)
-        self.loop.run_until_complete(asgi_task)
-        return self.response
-
-    async def receive(self) -> dict:
-        message = await self.app_queue.get()
-        return message
-
-    def put_message(self, message: ASGIMessage) -> None:
-        self.app_queue.put_nowait(message)
-
-
-@dataclass
-class ASGIHTTPCycle(ASGICycle):
-
-    body: bytes = b""
-
-    async def send(self, message: ASGIMessage) -> None:
-        if self.state is ASGICycleState.REQUEST:
-            if message["type"] != "http.response.start":
-                raise RuntimeError(
-                    f"Expected 'http.response.start', received: {message['type']}"
-                )
-
-            status_code = message["status"]
-            headers = {k: v for k, v in message.get("headers", [])}
-            self.response["statusCode"] = status_code
-            self.response["isBase64Encoded"] = self.binary
-            self.response["headers"] = {
-                k.decode(): v.decode() for k, v in headers.items()
-            }
-            self.state = ASGICycleState.RESPONSE
-
-        elif self.state is ASGICycleState.RESPONSE:
-            if message["type"] != "http.response.body":
-                raise RuntimeError(
-                    f"Expected 'http.response.body', received: {message['type']}"
-                )
-
-            body = message.get("body", b"")
-            more_body = message.get("more_body", False)
-
-            # The body must be completely read before returning the response.
-            self.body += body
-
-            if not more_body:
-                body = self.body
-                if self.binary:
-                    body = base64.b64encode(body)
-                self.response["body"] = body.decode()
-                self.put_message({"type": "http.disconnect"})
-
-
-@dataclass
-class ASGIWebSocketCycle(ASGICycle):
-
-    endpoint_url: str = None
-
-    async def send(self, message: ASGIMessage) -> None:
-        if self.state is not ASGICycleState.REQUEST:
-            if message["type"] in ("websocket.accept", "websocket.close"):
-                self.response["statusCode"] = 200
-                self.response["headers"] = {"content-type": "text/plain; charset=utf-8"}
-                self.response["body"] = "OK"
-                self.state = ASGICycleState.RESPONSE
-            else:
-                raise RuntimeError(
-                    f"Expected 'websocket.accept' or 'websocket.close', received: {message['type']}"
-                )
-        else:
-            bytes_data = message.get("bytes")
-            text_data = message.get("text")
-            data = text_data if bytes_data is None else bytes_data
-            if message["type"] == "websocket.send":
-                connections = get_connections()
-                items = connections.scan(ProjectionExpression="connectionId").get(
-                    "Items"
-                )
-                if items is None:
-                    self.response["statusCode"] = 500
-                    self.response["headers"] = {
-                        "content-type": "text/plain; charset=utf-8"
-                    }
-                    self.response["body"] = "Connection error"
-                else:
-                    self.response = send_to_connections(
-                        data=data,
-                        connections=connections,
-                        items=items,
-                        endpoint_url=self.endpoint_url,
-                    )
-
-                    # apigw_management = boto3.client(
-                    #     "apigatewaymanagementapi", endpoint_url=self.endpoint_url
-                    # )
-                    # for item in items:
-                    #     try:
-                    #         apigw_management.post_to_connection(
-                    #             ConnectionId=item["connectionId"], Data=data
-                    #         )
-                    #     except botocore.exceptions.ClientError as exc:
-                    #         status_code = exc.response.get("ResponseMetadata", {}).get(
-                    #             "HTTPStatusCode"
-                    #         )
-                    #         if status_code == 410:
-                    #             connections.delete_item(
-                    #                 Key={"connectionId": item["connectionId"]}
-                    #             )
-                    #         else:
-                    #             self.response["body"] = "Connection error"
-                    #             self.response["statusCode"] = 500
+from mangum.utils import get_logger, make_response, get_connections
+from mangum.types import ASGIScope, ASGIApp, AWSEvent, AWSContext, AWSResponse
+from mangum.asgi import ASGIHTTPCycle, ASGIWebSocketCycle
 
 
 @dataclass
